feat: Add polls tab with example question
This commit adds a new "polls" tab to the tab layout. The tab is represented by a new file "polls.tsx" which contains the necessary components and styles for displaying a poll question. The tab is also given a title of "polls" and an icon representing a podium.

Recent user commits:
- chore: Update app.json slug to "Polls"
- chore: Update app.json scheme to "Polls" and tab layout title to "Home Vote"

Recent repository commits:
- chore: Update app.json slug to "Polls"
- chore: Update app.json scheme to "Polls" and tab layout title to "Home Vote"
- chore: Update explore screen layout and styles
- Initial commit