feat: Add Stack.Screen for CreatePoll title and "New Poll" text
This commit adds a Stack.Screen component to the CreatePoll function in new.tsx. The Stack.Screen component includes options for setting the title of the screen to "CreatePoll". Additionally, a Text component with the text "New Poll" is added to the CreatePoll function.

Recent user commits:
- feat: Add header title, input fields for question and options in CreatePoll component
- chore: Remove empty text element in CreatePoll component

Recent repository commits:
- feat: Update polls tab to use FlatList for rendering
- refactor: Remove unused code for explore and not-found screens